requirements.txt
langchain
openai
langchain_google_genai
streamlit
langchain_core
google-generativeai
os
langchain-community
langchain-openai 
python-dotenv